+++ database/init_db.production/GHS_POP.sh
@@ -9,6 +9,9 @@ unzip GHS_POP_E2015_GLOBE_R2019A_4326_9ss_V1_0
 
 geotiff=GHS_POP_E2015_GLOBE_R2019A_4326_9ss_V1_0.tif
 
+psql --command \
+    "DROP TABLE IF EXISTS ghs_pop"
+
 raster2pgsql \
     -I \
     -M \