+++ database/scripts/create_dev_schema.sql
@@ -1,17 +1,14 @@
 /* */
-/* The purpose of this SQL script is to create a schema with datasets for development.*/
+/* The purpose of this SQL script is to create schemas with data for development and testing. */
+/* Data for testing is smaller. */
 /* The schema including its data can be dumped using pg_dump. */
-/* This is usefull to setup a local development database. */
+/* This is usefull to setup a local development or testing database. */
 /* It is used as basis for the development database inside a Docker container. */
 /* */
-SET search_path TO public, development;
+SET search_path TO public, development, test;
 
-DROP TABLE IF EXISTS development.regions;
 
-DROP TABLE IF EXISTS development.ghs_pop;
-
-DROP SCHEMA IF EXISTS development;
-
+/* Development */
 CREATE SCHEMA IF NOT EXISTS development;
 
 CREATE TABLE development.regions (
@@ -22,8 +19,6 @@ CREATE TABLE development.ghs_pop (
     LIKE public.ghs_pop INCLUDING INDEXES
 );
 
-
-/* Exclude big regions */
 INSERT INTO development.regions
 SELECT
     *
@@ -31,7 +26,7 @@ FROM
     public.regions;
 
 INSERT INTO development.ghs_pop
-SELECT
+SELECT DISTINCT ON (rid)
     rid,
     ST_Clip (rast, ST_Buffer (geom, 0.01), TRUE) AS rast
 FROM
@@ -39,7 +34,43 @@ FROM
     development.regions
 WHERE
     ST_Intersects (rast, geom)
-    AND ST_BandIsNoData (rast) = FALSE;
+    AND ST_BandIsNoData (rast) = FALSE
+    AND name NOT IN ('Bangladesh', 'Dominican Republic', 'Haiti', 'Myanmar',
+	'South Sudan');
 
 SELECT
     AddRasterConstraints ('development'::name, 'ghs_pop'::name, 'rast'::name);
+
+
+/* Testing */
+CREATE SCHEMA IF NOT EXISTS test;
+
+CREATE TABLE test.regions (
+    LIKE public.regions INCLUDING INDEXES
+);
+
+CREATE TABLE test.ghs_pop (
+    LIKE public.ghs_pop INCLUDING INDEXES
+);
+
+INSERT INTO test.regions
+SELECT
+    *
+FROM
+    public.regions
+WHERE
+    ogc_fid IN (1, 2, 3, 4, 8, 9, 14, 28, 31);
+
+INSERT INTO test.ghs_pop
+SELECT DISTINCT ON (rid)
+    rid,
+    ST_Clip (rast, ST_Buffer (geom, 0.01), TRUE) AS rast
+FROM
+    public.ghs_pop,
+    test.regions
+WHERE
+    ST_Intersects (rast, geom)
+    AND ST_BandIsNoData (rast) = FALSE;
+
+SELECT
+    AddRasterConstraints ('test'::name, 'ghs_pop'::name, 'rast'::name);