+++ website/scripts/get_regions.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+wget https://downloads.ohsome.org/OQT/regions.geojson